+++ app/CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.15] - 07/08/26
+
+## Added
+- Daily statistics on ome screen.
+- Pause/resume sync service when app is paused/resumed.
+
 ## [0.1.14] - 04/08/26
 
 ### Fixed
@@ -143,3 +149,5 @@ Initial beta app release
 [0.1.11]: https://github.com/ConcenTech/grobase/compare/v0.1.10...v0.1.11
 [0.1.12]: https://github.com/ConcenTech/grobase/compare/v0.1.11...v0.1.12
 [0.1.13]: https://github.com/ConcenTech/grobase/compare/v0.1.12...v0.1.13
+[0.1.14]: https://github.com/ConcenTech/grobase/compare/v0.1.13...v0.1.14
+[0.1.15]: https://github.com/ConcenTech/grobase/compare/v0.1.14...v0.1.15